move schema explorer to core-components
