+++ packages/core-components/src/lib/unsorted/ui/Explore/index.js
@@ -1,0 +1,1 @@
+export { default as SchemaExplorer } from './SchemaExplorer.svelte';
