+++ packages/core-components/src/lib/unsorted/ui/index.js
@@ -23,3 +23,4 @@ export * from './QueryViewerSupport';
 export * from './Tabs';
 export * from './TelemetryOptOut';
 export * from './VersionControl';
+export * from './Explore';
