+++ sites/example-project/src/pages/explore/schema/+page.svelte
@@ -1,112 +1,11 @@
 <script>
-	import { browser } from '$app/environment';
-	import { Icon } from '@steeze-ui/svelte-icon';
-	import { Abc, Calendar, _123, CircleHalf2, Table, Database } from '@steeze-ui/tabler-icons';
+	import { SchemaExplorer } from "@evidence-dev/core-components";
 	export let data;
-	let { __db: db } = data;
-
-	async function loadMetadata() {
-		if (!browser) return {};
-
-		const tables = await db.query(
-			`SELECT * FROM information_schema.tables WHERE table_catalog = 'memory' AND table_name != 'stats'`
-		);
-
-		const metadatas = {};
-
-		await Promise.all(
-			tables.map(async (t) => {
-				const columns = await db.query(
-					`SELECT * FROM information_schema.columns WHERE table_name = '${t.table_name}' AND table_schema = '${t.table_schema}'`
-				);
-
-				const sourceName = `${t.table_schema}.${t.table_name}`;
-
-				if (!metadatas[t.table_schema]) {
-					metadatas[t.table_schema] = {};
-				}
-
-				metadatas[t.table_schema][t.table_name] = { table: t, columns };
-			})
-		);
-
-		return metadatas;
-	}
-
-	let selectedTable = '';
-	let selectedSource = '';
 </script>
 
-<h1 class="text-xl">Project Schema</h1>
-<p>This page details the tables and columns that are currently loaded in your project.</p>
+<h1 class="markdown">Project Schema</h1>
+<p class="markdown">This page details the tables and columns that are currently loaded in your project.</p>
 
-{#await loadMetadata()}
-	Loading Schema Information...
-{:then metadata}
-	<section>
-		<div>
-			<h2 class="text-base font-normal font-mono my-2">Sources</h2>
-			<ul class="list-none m-0 p-0 flex flex-col gap-1">
-				{#each Object.entries(metadata) as [source, meta] (source)}
-					<li class="font-mono m-0 text-sm text-white">
-						<button
-							class="bg-gray-500 px-2 py-1 rounded font-bold flex w-full hover:bg-blue-500"
-							class:bg-blue-500={selectedSource === source}
-							on:click={() => {
-								selectedSource = selectedSource === source ? '' : source;
-								selectedTable = ''; // Reset selectedTable when source is clicked
-							}}
-						>
-							<Icon src={Database} class="text-white w-5 h-5 mr-1" />
-							{source}
-						</button>
-					</li>
-					{#if selectedSource === source}
-						<ul class="list-none m-0 flex flex-col gap-1">
-							{#each Object.entries(meta) as [name, tableMeta] (name)}
-								<li class="font-mono m-0 text-sm font-bold ml-3">
-									<button
-										class="bg-gray-300 px-2 py-1 rounded flex w-full hover:bg-blue-200"
-										class:bg-blue-300={selectedTable === tableMeta}
-										on:click={() => {
-											selectedTable = selectedTable === tableMeta ? '' : tableMeta;
-										}}
-									>
-										<Icon src={Table} class="text-gray-700 w-5 h-5 mr-1" />
-										{name}
-									</button>
-								</li>
-								{#if selectedTable === tableMeta}
-									<ul class="list-none m-0 flex flex-col gap-1">
-										{#each tableMeta.columns as column (column.column_name)}
-											<li class="font-mono text-sm ml-6 rounded flex flex-row">
-												<div class="flex px-2 py-1 rounded w-full hover:bg-blue-100">
-													{#if column.data_type === 'INT' || column.data_type === 'BIGINT' || column.data_type === 'SMALLINT' || column.data_type === 'TINYINT' || column.data_type === 'DOUBLE'}
-														<Icon src={_123} class="text-gray-700 w-5 h-5" />
-													{:else if column.data_type === 'DATE' || column.data_type === 'DATETIME' || column.data_type === 'TIMESTAMP'}
-														<Icon src={Calendar} class="text-gray-700 w-5 h-5" />
-													{:else if column.data_type === 'BOOLEAN'}
-														<Icon src={CircleHalf2} class="text-gray-700 w-5 h-5" />
-													{:else}
-														<Icon src={Abc} class="text-gray-700 w-5 h-5" />
-													{/if}
-													<div class="pl-2 lowercase">
-														<b>{column.column_name}</b>&nbsp; {column.data_type}
-													</div>
-												</div>
-											</li>
-										{/each}
-									</ul>
-								{/if}
-							{/each}
-						</ul>
-					{/if}
-				{/each}
-			</ul>
-		</div>
-	</section>
-{:catch e}
-	An error was encountered while loading project schema.
+<h2 class="markdown">Sources</h2>
 
-	<pre class="px-4 py-2 bg-red-800 text-white">{e.message}</pre>
-{/await}
+<SchemaExplorer data={data}/>
